Update 20533C_LAB_AK_09.md
Removed steps referencing the devices section of the CONFIGURATION page of the Adatum Azure Active Directory tenant. This section no longer appears in the Azure classic portal. The default settings allow Azure AD users to join devices to Azure AD, so the remainder of the lab still works as intended.